refactor: #52 variant와 size token 파일 분리
